Broke out handleUrlSubmit module and refactored main.js accordingly. #31
